Refactored text decorator into utils.
